requirements.txt
aiolimiter==1.1.0
anyio==3.7.1
autopep8==2.0.2
certifi==2023.5.7
charset-normalizer==3.2.0
colorama==0.4.6
dnspython==2.3.0
exceptiongroup==1.1.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
packaging==23.1
pluggy==1.2.0
pycodestyle==2.10.0
pymongo==4.4.1
pytest==7.4.0
python-dotenv==1.0.0
python-telegram-bot==20.4
requests==2.31.0
sniffio==1.3.0
tomli==2.0.1
urllib3==2.0.3
